Resolve conflicts: keep main's CancellationToken signatures,
re-apply our GraphQL query fixes (full schema fields, heatPumpStatus
accountNumber arg, heatPumpControllersAtLocation ID! type).

+++ OctopusCosyAnalyser.ApiService/Models/DailyCostRecord.cs
@@ -5,10 +5,10 @@ public class DailyCostRecord
     public int Id { get; set; }
     public string DeviceId { get; set; } = string.Empty;
     public DateOnly Date { get; set; }
-    public double TotalCostPence { get; set; }
-    public double TotalUsageKwh { get; set; }
-    public double AvgUnitRatePence { get; set; }
-    public double? StandingChargePence { get; set; }
+    public decimal TotalCostPence { get; set; }
+    public decimal TotalUsageKwh { get; set; }
+    public decimal AvgUnitRatePence { get; set; }
+    public decimal? StandingChargePence { get; set; }
     public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
     public DateTime UpdatedAt { get; set; } = DateTime.UtcNow;
 }
